requirements.txt
anyio==3.3.4
asgiref==3.4.1
attrs==21.2.0
autocfg==0.0.8
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
cycler==0.11.0
fastapi==0.70.0
gluoncv==0.10.4.post4
graphviz==0.8.4
h11==0.12.0
idna==3.3
iniconfig==1.1.1
kiwisolver==1.3.2
matplotlib==3.4.3
mxnet==1.8.0.post0
numpy==1.21.4
opencv-python==4.5.4.58
packaging==21.2
pandas==1.3.4
Pillow==8.4.0
pluggy==1.0.0
portalocker==2.3.2
py==1.11.0
pydantic==1.8.2
pyparsing==2.4.7
pytest==6.2.5
python-dateutil==2.8.2
python-multipart==0.0.5
pytz==2021.3
PyYAML==6.0
requests==2.26.0
scipy==1.7.2
six==1.16.0
sniffio==1.2.0
starlette==0.16.0
toml==0.10.2
tqdm==4.62.3
typing-extensions==3.10.0.2
urllib3==1.26.7
uvicorn==0.15.0
yacs==0.1.8
